+++ src/pages/OrderPage.jsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 
 import useOrderStore from '../hooks/useOrderStore';
 
-import Order from '../components/Order';
+import OrderForm from '../components/OrderForm';
 
 export default function OrderPage() {
   const orderStore = useOrderStore();
@@ -12,6 +12,6 @@ export default function OrderPage() {
   }, []);
 
   return (
-    <Order />
+    <OrderForm />
   );
 }
